refactor: Unify spanConfig add function
